+++ core/src/main/kotlin/diary/service/DiaryService.kt
@@ -25,7 +25,6 @@ import com.wafflestudio.snutt.notification.service.PushService
 import com.wafflestudio.snutt.semester.service.SemesterService
 import com.wafflestudio.snutt.timetables.repository.TimetableRepository
 import kotlinx.coroutines.flow.toList
-import org.slf4j.LoggerFactory
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
 import java.time.Instant
@@ -65,8 +64,6 @@ interface DiaryService {
     suspend fun addQuestion(request: DiaryAddQuestionRequestDto)
 
     suspend fun removeQuestion(questionId: String)
-
-    suspend fun sendNotifier()
 }
 
 @Service
@@ -80,12 +77,6 @@ class DiaryServiceImpl(
     private val pushService: PushService,
     private val cache: Cache,
 ) : DiaryService {
-    private val logger = LoggerFactory.getLogger(this::class.java)
-
-    companion object {
-        const val SAMPLE_RATE = 0.1
-    }
-
     override suspend fun generateQuestionnaire(
         userId: String,
         lectureId: String,
@@ -99,9 +90,12 @@ class DiaryServiceImpl(
             timetableRepository.findByUserIdAndYearAndSemesterAndIsPrimaryTrue(userId, lecture.year, lecture.semester)
                 ?: throw TimetableNotFoundException
 
-        val recentlySubmittedIds = diarySubmissionRepository.findAllByUserIdAndCreatedAtIsAfter(userId,
-            LocalDateTime.now().minusDays(1)).map { it.lectureId }
-        val nextLectureCandidates = userTimetable.lectures.filterNot { it.lectureId == lectureId || recentlySubmittedIds.contains(it.lectureId) }
+        val recentlySubmittedIds = diarySubmissionRepository.findAllByUserIdAndCreatedAtIsAfter(
+            userId,
+            LocalDateTime.now().minusDays(1)
+        ).map { it.lectureId }
+        val nextLectureCandidates =
+            userTimetable.lectures.filterNot { it.lectureId == lectureId || recentlySubmittedIds.contains(it.lectureId) }
         val nextLecture = nextLectureCandidates.randomOrNull()
 
         val questions =
@@ -214,46 +208,4 @@ class DiaryServiceImpl(
         question.active = false
         diaryQuestionRepository.save(question)
     }
-
-    @Scheduled(cron = "0 0 19 * * MON,WED,FRI", zone = "Asia/Seoul")
-    override suspend fun sendNotifier() {
-        val lockKey = CacheKey.LOCK_SEND_LECTURE_DIARY_NOTIFICATION.build()
-        cache.withLock(lockKey) {
-            try {
-                val currentTime = Instant.now()
-                val (currentYear, currentSemester) =
-                    semesterService.getCurrentYearAndSemester(currentTime) ?: run {
-                        logger.debug("현재 진행 중인 학기가 없습니다.")
-                        return@withLock
-                    }
-                val sampledUserIdPrimaryTimetableMap =
-                    timetableRepository
-                        .samplePrimaryOfRateByYearAndSemester(SAMPLE_RATE, currentYear, currentSemester)
-                        .toList()
-                        .filter { it.lectures.size > 2 }
-                        .associateBy { it.userId }
-                val targetedPushMessages =
-                    sampledUserIdPrimaryTimetableMap
-                        .mapNotNull {
-                            val targetLecture = it.value.lectures.randomOrNull() ?: return@mapNotNull null
-                            it.key to buildPushMessage(targetLecture.lectureId!!, targetLecture.courseTitle)
-                        }.toMap()
-
-                pushService.sendTargetPushes(targetedPushMessages, PushPreferenceType.DIARY)
-            } catch (e: Exception) {
-                logger.error("강의 일기장 알림 전송 중 문제 발생", e)
-            }
-        }
-    }
-
-    private fun buildPushMessage(
-        lectureId: String,
-        courseTitle: String,
-    ): PushMessage =
-        PushMessage(
-            title = "이번주 강의일기를 작성해보세요.",
-            body = "최근 수강한 <$courseTitle> 강의에 대한 강의일기를 작성해보세요.\uD83D\uDCD4 ",
-            urlScheme = DeeplinkType.DIARY.build(lectureId),
-            isUrgentOnAndroid = false,
-        )
 }
